#19 Add a preview page button in the actions blocks
On front verification is made based on a 'token', we can't rely on a role cause admin and front are not on the same firewall and in general admin and front are not on the same host.

Save button in the admin as been moved to the seo panel because it's the only thing that it actually save.